refactor: migrate loading-spinner
