+++ packages/components/loading-spinner/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './loading-spinner';
+
+export { default as version } from './version';
