+++ packages/components/loading-spinner/src/loading-spinner.js
