+++ packages/components/loading-spinner/src/loading-spinner.spec.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render } from '../../test-utils';
+import { render } from '../../../../src/test-utils';
 import LoadingSpinner from './loading-spinner';
 
 it('should render children', () => {
