+++ packages/components/loading-spinner/src/loading-spinner.story.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs, select, text } from '@storybook/addon-knobs/react';
-import Section from '../../../.storybook/decorators/section';
-import Readme from './README.md';
+import Section from '../../../../.storybook/decorators/section';
+import Readme from '../README.md';
 import LoadingSpinner from './loading-spinner';
 
 storiesOf('Components|Loading', module)
