+++ packages/components/loading-spinner/src/loading-spinner.visualspec.js
