+++ packages/components/loading-spinner/src/version.js
