--- src/components/loading-spinner/index.js
@@ -1,1 +1,0 @@
-export { default } from './loading-spinner';
